+++ src/Phinx/Console/Command/Rollback.php
@@ -47,6 +47,8 @@ class Rollback extends AbstractCommand
              ->setDescription('Rollback the last or to a specific migration')
              ->addOption('--target', '-t', InputOption::VALUE_REQUIRED, 'The version number to rollback to')
              ->addOption('--date', '-d', InputOption::VALUE_REQUIRED, 'The date to rollback to')
+             ->addOption('--force', '-f', InputOption::VALUE_NONE, 'Force rollback to ignore breakpoints')
+             ->addOption('--start-time', '-s', InputOption::VALUE_NONE, 'Order the migrations by their start time (ie. execution time)')
              ->setHelp(
 <<<EOT
 The <info>rollback</info> command reverts the last migration, or optionally up to a specific version
@@ -55,6 +57,18 @@ The <info>rollback</info> command reverts the last migration, or optionally up t
 <info>phinx rollback -e development -t 20111018185412</info>
 <info>phinx rollback -e development -d 20111018</info>
 <info>phinx rollback -e development -v</info>
+<info>phinx rollback -e development -t 20111018185412 -f</info>
+
+If you have a breakpoint set, then you can rollback to target 0 and the rollbacks will stop at the breakpoint.
+<info>phinx rollback -e development -t 0 </info>
+
+The <info>-s|--start-time</info> option can be used to allow rollbacking the last executed migration (as opposed to the last created one):
+
+<info>phinx rollback -s</info>
+
+It can also be combined with the <info>-d|--date</info> option to rollback to a certain date using the migration start times to order them:
+
+<info>phinx rollback -s -d 20111018</info>
 
 EOT
              );
@@ -74,15 +88,19 @@ EOT
         $environment = $input->getOption('environment');
         $version     = $input->getOption('target');
         $date        = $input->getOption('date');
+        $force       = !!$input->getOption('force');
+        $startTime   = $input->getOption('start-time');
+
+        $config = $this->getConfig();
 
         if (null === $environment) {
-            $environment = $this->getConfig()->getDefaultEnvironment();
+            $environment = $config->getDefaultEnvironment();
             $output->writeln('<comment>warning</comment> no environment specified, defaulting to: ' . $environment);
         } else {
             $output->writeln('<info>using environment</info> ' . $environment);
         }
 
-        $envOptions = $this->getConfig()->getEnvironment($environment);
+        $envOptions = $config->getEnvironment($environment);
         if (isset($envOptions['adapter'])) {
             $output->writeln('<info>using adapter</info> ' . $envOptions['adapter']);
         }
@@ -94,13 +112,35 @@ EOT
         if (isset($envOptions['name'])) {
             $output->writeln('<info>using database</info> ' . $envOptions['name']);
         }
+        
+        if (isset($config['rollback_order'])) {
+            if ('start-time' === $config['rollback_order']) {
+                $startTime = true;
+            }
+            else if ('creation' !== $config['rollback_order']) {
+                $output->writeln('<error>Invalid rollback_order environment option: ' 
+                    . $config['rollback_order'] . '. Valid values: creation or start-time</error>');
+            }
+        }
+
+        if ($startTime) {
+            $output->writeln('<info>ordering by </info>start time');
+        }
 
         // rollback the specified environment
         $start = microtime(true);
         if (null !== $date) {
-            $this->getManager()->rollbackToDateTime($environment, new \DateTime($date));
+            if (true === $startTime) {
+                $this->getManager()->rollbackToDateTimeByStartTime($environment, new \DateTime($date), $force);
+            } else {
+                $this->getManager()->rollbackToDateTime($environment, new \DateTime($date), $force);
+            }
         } else {
-            $this->getManager()->rollback($environment, $version);
+            if (true === $startTime) {
+                $this->getManager()->rollbackByStartTime($environment, $version, $force);
+            } else {
+                $this->getManager()->rollback($environment, $version, $force);
+            }
         }
         $end = microtime(true);