+++ tests/Phinx/Console/Command/StatusTest.php
@@ -2,6 +2,13 @@
 
 namespace Test\Phinx\Console\Command;
 
+use Phinx\Config\ConfigInterface;
+use Phinx\Console\PhinxApplication;
+use Phinx\Migration\Manager;
+use PHPUnit_Framework_MockObject_MockObject;
+use Symfony\Component\Console\Input\ArrayInput;
+use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Tester\CommandTester;
 use Symfony\Component\Console\Output\StreamOutput;
 use Phinx\Config\Config;
@@ -9,7 +16,20 @@ use Phinx\Console\Command\Status;
 
 class StatusTest extends \PHPUnit_Framework_TestCase
 {
+    /**
+     * @var ConfigInterface|array
+     */
     protected $config = array();
+
+    /**
+     * @var InputInterface $input
+     */
+    protected $input;
+
+    /**
+     * @var OutputInterface $output
+     */
+    protected $output;
 
     protected function setUp()
     {
@@ -30,20 +50,22 @@ class StatusTest extends \PHPUnit_Framework_TestCase
                 )
             )
         ));
+
+        $this->input = new ArrayInput([]);
+        $this->output = new StreamOutput(fopen('php://memory', 'a', false));
     }
 
     public function testExecute()
     {
-        $application = new \Phinx\Console\PhinxApplication('testing');
+        $application = new PhinxApplication('testing');
         $application->add(new Status());
 
-        // setup dependencies
-        $output = new StreamOutput(fopen('php://memory', 'a', false));
-
+        /** @var Status $command */
         $command = $application->find('status');
 
         // mock the manager class
-        $managerStub = $this->getMock('\Phinx\Migration\Manager', array(), array($this->config, $output));
+        /** @var Manager|PHPUnit_Framework_MockObject_MockObject $managerStub */
+        $managerStub = $this->getMock('\Phinx\Migration\Manager', array(), array($this->config, $this->input, $this->output));
         $managerStub->expects($this->once())
                     ->method('printStatus')
                     ->will($this->returnValue(0));
@@ -60,16 +82,15 @@ class StatusTest extends \PHPUnit_Framework_TestCase
 
     public function testExecuteWithEnvironmentOption()
     {
-        $application = new \Phinx\Console\PhinxApplication('testing');
+        $application = new PhinxApplication('testing');
         $application->add(new Status());
 
-        // setup dependencies
-        $output = new StreamOutput(fopen('php://memory', 'a', false));
-
+        /** @var Status $command */
         $command = $application->find('status');
 
         // mock the manager class
-        $managerStub = $this->getMock('\Phinx\Migration\Manager', array(), array($this->config, $output));
+        /** @var Manager|PHPUnit_Framework_MockObject_MockObject $managerStub */
+        $managerStub = $this->getMock('\Phinx\Migration\Manager', array(), array($this->config, $this->input, $this->output));
         $managerStub->expects($this->once())
                     ->method('printStatus')
                     ->will($this->returnValue(0));
@@ -85,16 +106,15 @@ class StatusTest extends \PHPUnit_Framework_TestCase
 
     public function testFormatSpecified()
     {
-        $application = new \Phinx\Console\PhinxApplication('testing');
+        $application = new PhinxApplication('testing');
         $application->add(new Status());
 
-        // setup dependencies
-        $output = new StreamOutput(fopen('php://memory', 'a', false));
-
+        /** @var Status $command */
         $command = $application->find('status');
 
         // mock the manager class
-        $managerStub = $this->getMock('\Phinx\Migration\Manager', array(), array($this->config, $output));
+        /** @var Manager|PHPUnit_Framework_MockObject_MockObject $managerStub */
+        $managerStub = $this->getMock('\Phinx\Migration\Manager', array(), array($this->config, $this->input, $this->output));
         $managerStub->expects($this->once())
                     ->method('printStatus')
                     ->will($this->returnValue(0));