+++ tests/Phinx/Console/PhinxApplicationTest.php
@@ -2,6 +2,7 @@
 
 namespace Test\Phinx\Console;
 
+use Phinx\Console\Command\AbstractCommand;
 use Symfony\Component\Console\Tester\ApplicationTester;
 use Phinx\Console\PhinxApplication;
 
@@ -9,10 +10,13 @@ class PhinxApplicationTest extends \PHPUnit_Framework_TestCase
 {
     /**
      * @dataProvider provider
+     *
+     * @param AbstractCommand $command
+     * @param $result
      */
     public function testRun($command, $result)
     {
-        $app = new \Phinx\Console\PhinxApplication('testing');
+        $app = new PhinxApplication('testing');
         $app->setAutoExit(false); // Set autoExit to false when testing
         $app->setCatchExceptions(false);