+++ CHANGELOG.md
@@ -1,7 +1,25 @@
 # Version History
 
-**0.5.x-dev** (****)
+**0.6.x-dev** (***)
 
+* Documentation cleanup
+* Bug fixes
+  [#884](https://github.com/robmorgan/phinx/pull/884) Only rollback 1 migration when only 2 migrations exist
+
+**0.6.2** (Thursday, 23rd June 2016)
+
+* Fix breakpoint support for Postgres
+* HHVM now passes all tests
+
+**0.6.1** (Tuesday, 21st June 2016)
+
+* Fix rollback when only 1 migration
+
+**0.6.0** (Tuesday, 21st June 2016)
+
+* Backward incompatibility - see [UPGRADE_0.6](UPGRADE_0.6.md) document
+* Introduce Input and Output access to migrations and template creation
+* New breakpoint command
 * Moved version history to this CHANGELOG.md document
 * More tests